Files

+++ server/src/generate_test_data.py
@@ -147,7 +147,8 @@ async def main():
     await db[DbName.QUESTION.value].drop()
     await db[DbName.COMMENT.value].drop()
     await db[DbName.SIMULATION.value].drop()
-    shutil.rmtree("/server/media/TEST")
+    if os.path.exists("/server/media/TEST"):
+        shutil.rmtree("/server/media/TEST")
 
     quiz_xml = open(QUIZ_FILEPATH, "rb")
     quiz_id = ObjectId()
